refactor : QrCertificationFragment로 네이밍 변경
