--- app/src/main/java/com/minionz/qrna/view/ui/dashboard/DashboardViewModel.kt
@@ -1,9 +1,0 @@
-package com.minionz.qrna.view.ui.dashboard
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class DashboardViewModel : ViewModel() {
-
-}
