+++ app/src/main/java/com/minionz/qrna/view/ui/qrcertification/QrCertificationViewModel.kt
@@ -1,0 +1,7 @@
+package com.minionz.qrna.view.ui.qrcertification
+
+import androidx.lifecycle.ViewModel
+
+class QrCertificationViewModel : ViewModel() {
+
+}
