+++ app/src/main/res/values/strings.xml
@@ -11,7 +11,7 @@
     <string name="sign_up_cancel">Cancel</string>
     <string name="title_activity_main">MainActivity</string>
     <string name="title_home">홈</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_qr_certification">QR 인증</string>
     <string name="title_my_info">내 정보</string>
     <string name="sign_up_address">주소</string>
     <string name="sign_in_button_text">SignIn</string>
